Create a new C# project and create a program that assigns integer values to variables.
Create a new C# project and create a program that assigns integer values
to variables. Be sure that each value is stored in the correct variable
type (try to find the most suitable variable type in order to save
memory). Finally, you need to print all variables to the console.